metrics-recorder: Split off `helpers` module
